// src/types/schema.ts
export interface components {
    schemas: {
// ...
        EventPreview: {
            /** Format: uuid */
            id?: string;
            title?: string;
            coverImageUrl?: string;
            /** Format: date-time */
            startAt?: string;
            /** Format: date-time */
            endAt?: string;
            locationName?: string;
            status?: components["schemas"]["EnumPublicStatus"];
            organizers?: components["schemas"]["AssociationSimple"][];
            /** @description Dacă true, utilizatorul curent (din token) este deja înscris. */
            isRegistered?: boolean;
        };
// ...
        EventFilterQuery: {
            /** @default 1 */
            page: number;
            /** @default 20 */
            limit: number;
            /** @description Căutare text (titlu, descriere) */
            search?: string;
            /** @description Lista tipurilor selectate (ex: [SOCIAL, SPORTS]) */
            types?: components["schemas"]["EnumEventType"][];
            /**
             * @description Flag pentru logica de filtrare a tipurilor: - FALSE (Default): Returnează evenimentele care sunt de tip A **SAU** B. - TRUE: Returnează evenimentele care sunt și A **ȘI** B.
             * @default false
             */
            matchAllTypes: boolean;
            /** @description Selectare asociații specifice (ex: doar evenimentele ASUS și OSF) */
            associationIds?: string[];
            /** @description Filtrare după natura organizatorului */
            organizerTypes?: components["schemas"]["EnumOrganizerType"][];
            /** @description Selectare multiplă (ex: IN_CAMPUS + IN_CITY) */
            locationTypes?: components["schemas"]["EnumLocationType"][];
            /** @description Lista facultăților selectate (ex: ['FIESC', 'FEAA']) */
            faculties?: string[];
            /** @description Lista departamentelor selectate */
            departments?: string[];
            /** Format: date-time */
            dateFrom?: string;
            /** Format: date-time */
            dateTo?: string;
            /** @description [PROTEJAT] Permite filtrarea după starea tehnică. - Studenții: Ignorat (backend-ul forțează PUBLISHED). - Organizatori: Pot cere DRAFT, PENDING, REJECTED pentru asociația lor. - Adminii: Pot cere PENDING pentru a vedea coada de aprobare. */
            status?: components["schemas"]["EnumEventStatus"][];
        };
        AdminReport: {
            totalEvents?: number;
            totalRegistrations?: number;
            attendanceRate?: number;
            topAssociations?: string[];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}